Always endorse with the latest finalized block
